added slide functionality
